+++ lib/Crypto_Cli.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 require_relative 'crypto_cli/version'
-require_relative '../config/environment.rb'
-
+# require_relative '../config/environment.rb'
+# Contain App Version module
 module Crypto
+  puts "|===========================Crypto Version #{VERSION}===============================|"
 end
