--- lib/Crypto_Cli/tiker.rb
@@ -1,17 +1,0 @@
-class Ticker 
-  attr_accessor :name, :cmc_rank, :symbol, :price
-  @@all = []
-  
-  def initialize(name, cmc_rank, symbol, price)
-      @cmc_rank = cmc_rank
-      @name = name
-      @symbol = symbol
-      @price = price
-      @@all << self
-    end
-  end
-
-  def self.all 
-    @@all
-  end
-end
